Adding new chat button in Sidebar and moving Sidebar
